+++ packages/manager/cypress/support/api/domains.ts
@@ -1,8 +1,4 @@
-import {
-  getAll,
-  deleteById,
-  apiCheckErrors,
-} from './common';
+import { getAll, deleteById, apiCheckErrors } from './common';
 import { isTestLabel } from 'support/api/common';
 import { randomDomainName } from 'support/util/random';