+++ packages/manager/cypress/support/ui/events.ts
@@ -1,12 +1,8 @@
-// asserts text in popup toast notifications
-export const assertToast = (message: string, elementNumber: number = 1) => {
-  cy.get('[data-qa-toast]')
-    .then((elements) => {
-      if (elementNumber > 1) {
-        return elements[elementNumber - 1];
-      }
-    })
-    .within((_element) =>
-      cy.findByText(message, { exact: false }).should('be.visible')
-    );
+/**
+ * Asserts that a toast notification with the given message is visible.
+ *
+ * @param message - Message that should be displayed on toast notification.
+ */
+export const assertToast = (message: string) => {
+  cy.contains('[data-qa-toast]', message).should('be.visible');
 };