+++ packages/manager/src/components/NotificationDrawer/index.ts
@@ -1,6 +1,0 @@
-import NotificationDrawer, {
-  Props as _NotificationDrawerProps,
-} from './NotificationDrawer';
-// eslint-disable-next-line @typescript-eslint/no-empty-interface
-export interface NotificationDrawerProps extends _NotificationDrawerProps {}
-export default NotificationDrawer;