+++ packages/manager/src/features/Databases/DatabaseCreate/DatabaseCreate.tsx
@@ -25,6 +25,7 @@ import Paper from 'src/components/core/Paper';
 import RadioGroup from 'src/components/core/RadioGroup';
 import { makeStyles, Theme } from 'src/components/core/styles';
 import Typography from 'src/components/core/Typography';
+import DismissibleBanner from 'src/components/DismissibleBanner';
 import SingleValue from 'src/components/EnhancedSelect/components/SingleValue';
 import Select, { Item } from 'src/components/EnhancedSelect/Select';
 import RegionSelect from 'src/components/EnhancedSelect/variants/RegionSelect';
@@ -35,6 +36,7 @@ import Link from 'src/components/Link';
 import MultipleIPInput from 'src/components/MultipleIPInput';
 import Notice from 'src/components/Notice';
 import Radio from 'src/components/Radio';
+import { regionHelperText } from 'src/components/SelectRegionPanel/SelectRegionPanel';
 import TextField from 'src/components/TextField';
 import { databaseEngineMap } from 'src/features/Databases/DatabaseLanding/DatabaseRow';
 import { enforceIPMasks } from 'src/features/Firewalls/FirewallDetail/Rules/FirewallRuleDrawer';
@@ -55,7 +57,7 @@ import {
   validateIPs,
 } from 'src/utilities/ipUtils';
 import scrollErrorIntoView from 'src/utilities/scrollErrorIntoView';
-import DismissibleBanner from 'src/components/DismissibleBanner';
+import useFlags from 'src/hooks/useFlags';
 
 const useStyles = makeStyles((theme: Theme) => ({
   formControlLabel: {
@@ -119,7 +121,6 @@ const useStyles = makeStyles((theme: Theme) => ({
     },
   },
   notice: {
-    borderColor: theme.color.green,
     fontSize: 15,
     lineHeight: '18px',
   },
@@ -187,6 +188,7 @@ interface NodePricing {
 const DatabaseCreate: React.FC<{}> = () => {
   const classes = useStyles();
   const history = useHistory();
+  const flags = useFlags();
 
   const {
     data: regionsData,
@@ -417,20 +419,22 @@ const DatabaseCreate: React.FC<{}> = () => {
 
   return (
     <form onSubmit={handleSubmit}>
-      <DismissibleBanner
-        preferenceKey="dbaas-open-beta-notice"
-        productInformationIndicator
-      >
-        <Typography>
-          Managed Database for MySQL is available in a free, open beta period.
-          This is a beta environment and should not be used to support
-          production workloads. Review the{' '}
-          <Link to="https://www.linode.com/legal-eatp">
-            Early Adopter Program SLA
-          </Link>
-          .
-        </Typography>
-      </DismissibleBanner>
+      {flags.databaseBeta ? (
+        <DismissibleBanner
+          preferenceKey="dbaas-open-beta-notice"
+          productInformationIndicator
+        >
+          <Typography>
+            Managed Database for MySQL is available in a free, open beta period
+            until May 2nd, 2022. This is a beta environment and should not be
+            used to support production workloads. Review the{' '}
+            <Link to="https://www.linode.com/legal-eatp">
+              Early Adopter Program SLA
+            </Link>
+            .
+          </Typography>
+        </DismissibleBanner>
+      ) : null}
       <BreadCrumb
         labelTitle="Create"
         pathname={location.pathname}
@@ -488,18 +492,7 @@ const DatabaseCreate: React.FC<{}> = () => {
             regions={regionsData}
             selectedID={values.region}
           />
-          <Typography style={{ marginTop: 8 }}>
-            <a
-              target="_blank"
-              aria-describedby="external-site"
-              rel="noopener noreferrer"
-              href="https://www.linode.com/speed-test/"
-            >
-              Use our speedtest page
-            </a>
-            {` `}
-            to find the best region for your current location.
-          </Typography>
+          <div style={{ marginTop: 8 }}>{regionHelperText()}</div>
         </Grid>
         <Divider spacingTop={38} spacingBottom={12} />
         <Grid item>
@@ -556,14 +549,19 @@ const DatabaseCreate: React.FC<{}> = () => {
             </RadioGroup>
           </FormControl>
           <Grid item xs={12} md={8}>
-            <Notice success className={classes.notice}>
-              <strong>
-                Notice: There is no charge for database clusters during beta.
-              </strong>{' '}
-              You will be notified before the beta period ends and database
-              clusters are subject to charges.{' '}
-              <Link to="https://www.linode.com/pricing/">View pricing</Link>.
-            </Notice>
+            {flags.databaseBeta ? (
+              <Notice informational className={classes.notice}>
+                <strong>
+                  Notice: There is no charge for database clusters during beta.
+                </strong>{' '}
+                Database clusters will be subject to charges when the beta
+                period ends on May 2nd, 2022.{' '}
+                <Link to="https://www.linode.com/pricing/#databases">
+                  View pricing
+                </Link>
+                .
+              </Notice>
+            ) : undefined}
           </Grid>
         </Grid>
         <Divider spacingTop={26} spacingBottom={12} />