+++ packages/manager/src/features/Longview/LongviewDetail/DetailTabs/TopProcesses.test.tsx
@@ -53,7 +53,7 @@ describe('Top Processes', () => {
           />
         )
       );
-      getAllByTestId('skeletonCol');
+      getAllByTestId('table-row-loading');
     });
 
     it('renders error state', () => {