+++ packages/manager/src/features/NotificationCenter/index.tsx
@@ -1,2 +1,2 @@
 export { default as Events } from './Events';
-export { default as NotificationDrawer } from 'src/components/NotificationDrawer';
+export { default as NotificationMenu } from 'src/components/NotificationMenu';