+++ packages/manager/src/features/ObjectStorage/BucketDetail/ObjectTableContent.tsx
@@ -6,7 +6,7 @@ import { APIError } from '@linode/api-v4/lib/types';
 import * as React from 'react';
 import TableRowEmptyState from 'src/components/TableRowEmptyState';
 import TableRowError from 'src/components/TableRowError';
-import TableRowLoading from 'src/components/TableRowLoading';
+import { TableRowLoading } from 'src/components/TableRowLoading/TableRowLoading';
 import { useWindowDimensions } from 'src/hooks/useWindowDimensions';
 import { truncateEnd, truncateMiddle } from 'src/utilities/truncate';
 import { displayName, isFolder } from '../utilities';
@@ -41,7 +41,7 @@ const ObjectTableContent: React.FC<Props> = (props) => {
   const { width } = useWindowDimensions();
 
   if (isFetching && !isFetchingNextPage) {
-    return <TableRowLoading colSpan={4} widths={[20]} />;
+    return <TableRowLoading columns={4} responsive={{ 2: { smDown: true } }} />;
   }
 
   if (error) {
@@ -107,7 +107,9 @@ const ObjectTableContent: React.FC<Props> = (props) => {
           );
         });
       })}
-      {isFetchingNextPage && <TableRowLoading colSpan={12} transparent />}
+      {isFetchingNextPage ? (
+        <TableRowLoading columns={4} responsive={{ 2: { smDown: true } }} />
+      ) : null}
     </>
   );
 };