+++ packages/manager/src/features/linodes/LinodesCreate/LinodeCreate.tsx
@@ -581,7 +581,6 @@ export class LinodeCreate extends React.PureComponent<
               regions={regionsData!}
               handleSelection={this.props.updateRegionID}
               selectedID={this.props.selectedRegionID}
-              copy="Determine the best location for your Linode."
               updateFor={[this.props.selectedRegionID, regionsData, errors]}
               disabled={userCannotCreateLinode}
               helperText={this.props.regionHelperText}